Refactor AuthForms and AuthorizeForm for improved Steam integration
This commit updates the AuthForms and AuthorizeForm components to enhance
the user experience by resolving navigation issues and refining the page title updates.
The AuthorizeForm now uses the useRevalidator hook to ensure that Steam account
linking status is updated in real-time, addressing intermittent failures in session management.
Additionally, initial state setup for CloseInterceptReason is set to None to prevent
navigation interruptions when the AuthorizeForm component mounts.
